+++ http/src/main/scala/com/twitter/finatra/http/HttpHeaders.scala
@@ -1,8 +1,10 @@
 package com.twitter.finatra.http
 
 import com.twitter.finagle.http.Response
+import java.util.Locale
 import org.joda.time.format.DateTimeFormat
 import org.joda.time.{DateTime, DateTimeZone}
+
 
 object HttpHeaders {
 
@@ -48,7 +50,12 @@ object HttpHeaders {
    * RFC 7231 Date Format
    * @see <a href="http://tools.ietf.org/html/rfc7231#section-7.1.1.1">RFC 7231 Section 7.1.1.1</a>
    */
-  val RFC7231DateFormat = "EEE, dd MMM yyyy HH:mm:ss zzz"
+  val RFC7231DateFormat = "EEE, dd MMM yyyy HH:mm:ss 'GMT'"
+
+  val DateFormat = DateTimeFormat
+    .forPattern(RFC7231DateFormat)
+    .withZone(GMT)
+    .withLocale(Locale.US)
 
   /**
    * Set the given date under the given header name after formatting it as a string
@@ -58,8 +65,7 @@ object HttpHeaders {
    * @param date - the value to format and set as the header value
    */
   def setDate(response: Response, header: String, date: DateTime) {
-    val dateFormat = DateTimeFormat.forPattern(RFC7231DateFormat).withZone(GMT)
-    set(response, header, dateFormat.print(date))
+    set(response, header, DateFormat.print(date))
   }
 
   /**