+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file. Note that `
 
 ### Changed
 
+* finatra-http: fix HttpHeaders's Date locale problem. ``RB_ID=843966``
 * inject-thrift: Address issues with com.twitter.inject.exceptions.PossiblyRetryable. PossiblyRetryable does not correctly
   determine what is retryable. Updated to correct the logic for better default retry utility. ``RB_ID=843428``
 * finatra: finatra: Move com.twitter.finatra.annotations.Flag|FlagImpl to com.twitter.inject.annotations.Flag|FlagImpl. ``RB_ID=843383``